requirements.txt
torch
torchvision
pillow
streamlit
opencv-python
numpy